stabilization

+++ scripts/tools/verify-setup-scripts/verify-setup-scripts.sh
@@ -2,16 +2,15 @@
 
 set -e # exit on errors
 base_dir="$( cd "$(dirname "$0")" ; pwd -P )"
-
 source $base_dir/../../shared/bashHelpers.sh
 
 tmp_dir="/tmp/setuptests" # it is repeated on line 19
 setup_dir=$base_dir/../../../scripts/setup
-
 testcases_json=`cat ${base_dir}/setup-testcases.json`
 testcases_length=`echo $testcases_json | jq '. | length'`
 
 echoe "Testcases found: $testcases_length"
+
 if [ ! -d $tmp_dir ]; then
   echoe "Creating temp dir: $tmp_dir"
   mkdir -p $tmp_dir
@@ -24,11 +23,23 @@ fi
 declare -a pids
 
 killpids() {
-  for pid in pids[@]; do
-    echo "Killing $pid"
-    kill -9 $pid
+  for pid in "${pids[@]}"; do
+    # ps -p returns a header line and then the processes, so we count the lines with wc -l
+    # if there is more than 1 line, the process exists
+    if [ `ps -p $pid | wc -l` -gt 1 ]; then
+      echo "Killing $pid"
+      kill -9 $pid
+    fi
   done
 }
+
+# for ((i=0; i<$testcases_length; i++)); do
+#   sleep 10 &
+#   pid=$!
+#   pids+=($pid)
+#   killpids
+# done
+# exit 0
 
 for ((i=0; i<$testcases_length; i++)); do
   if [ $i -gt 0 ]; then
@@ -46,7 +57,10 @@ for ((i=0; i<$testcases_length; i++)); do
   test_baseurl=$(_jq '.baseurl')
   test_healthyurl=$(_jq '.healthyurl')
   test_waitafterhealthy=$(_jq '.waitafterhealthy')
+  test_webserverport=$(_jq '.webserverport')
   
+  killWebserver $test_webserverport
+
   cd $tmp_dir
   echoe "Running installation for ${test_name}"
   $setup_dir/$test_script $test_folder &
@@ -58,22 +72,31 @@ for ((i=0; i<$testcases_length; i++)); do
     echo "Testing $test_healthyurl"
     waitForServer $test_healthyurl 200
   } || {
+    echo "waitForServer failed"
     killpids
+    killWebserver $test_webserverport
+    exit 1
   }
 
   echo "Sleeping for $test_waitafterhealthy"
   sleep $test_waitafterhealthy # server is online a bit earlier most of the time (eg. with npm start)
   cd $base_dir/e2e
   {
-    $setup_dir/../node_modules/.bin/cypress run --config baseUrl=$test_baseurl
+    $setup_dir/../node_modules/.bin/cypress run --browser chrome --config baseUrl=$test_baseurl
   } || {
-    echoe "E2E exception: @ $__EXCEPTION_LINE__"
+    echoe "E2E failed"
     killpids
+    killWebserver $test_webserverport
+    exit 1
   }
 
+  echoe "Tear down"
+  killpids
+  killWebserver $test_webserverport
+
   if [ -d "${tmp_dir}/${test_folder}" ]; then
+    echoe "Removing testfolder ${test_folder}"
     rm -rf ${tmp_dir}/${test_folder}
   fi
-  killpids
   echoe "Successfully tested ${test_name}"
 done
